+++ examples/mixture/plot_gmm_sin.py
@@ -3,15 +3,16 @@
 Gaussian Mixture Model Sine Curve
 =================================
 
-This example highlights the advantages of the Dirichlet Process:
-complexity control and dealing with sparse data. The dataset is formed
-by 100 points loosely spaced following a noisy sine curve. The fit by
-the GMM class, using the expectation-maximization algorithm to fit a
-mixture of 10 Gaussian components, finds too-small components and very
-little structure. The fits by the Dirichlet process, however, show
-that the model can either learn a global structure for the data (small
-alpha) or easily interpolate to finding relevant local structure
-(large alpha), never falling into the problems shown by the GMM class.
+This example highlights the advantages of the Dirichlet Process: complexity
+control and dealing with sparse data. The dataset is formed by 100 points
+loosely spaced following a noisy sine curve. The fit by the ``GaussianMixture``
+class, using the expectation-maximization algorithm to fit a mixture of 10
+Gaussian components, finds too-small components and very little structure. The
+fits obtained by the ``BayesianGaussianMixture`` class with a Dirichlet
+process prior, however, show that the model can either learn a global structure
+for the data (small ``weight_concentration_prior``) or easily interpolate to
+finding relevant local structure (large ``weight_concentration_prior``), never
+falling into the problems shown by the ``GaussianMixture`` class.
 """
 
 import itertools
@@ -22,6 +23,8 @@ import matplotlib.pyplot as plt
 import matplotlib as mpl
 
 from sklearn import mixture
+
+print(__doc__)
 
 color_iter = itertools.cycle(['navy', 'c', 'cornflowerblue', 'gold',
                               'darkorange'])
@@ -73,13 +76,8 @@ def plot_samples(X, Y, n_components, index, title):
     plt.yticks(())
 
 
-plt.figure(figsize=(10, 10))
-plt.subplots_adjust(bottom=.04, top=0.95, hspace=.2, wspace=.05,
-                    left=.03, right=.97)
-
-# Number of samples per component
+# Parameters
 n_samples = 100
-
 
 # Generate random sample following a sine curve
 np.random.seed(0)
@@ -91,39 +89,42 @@ for i in range(X.shape[0]):
     X[i, 0] = x + np.random.normal(0, 0.1)
     X[i, 1] = 3. * (np.sin(x) + np.random.normal(0, .2))
 
+plt.figure(figsize=(10, 10))
+plt.subplots_adjust(bottom=.04, top=0.95, hspace=.2, wspace=.05,
+                    left=.03, right=.97)
+
 # Fit a Gaussian mixture with EM using ten components
 gmm = mixture.GaussianMixture(n_components=10, covariance_type='full',
                               max_iter=100).fit(X)
 plot_results(X, gmm.predict(X), gmm.means_, gmm.covariances_, 0,
              'Expectation-maximization')
 
-# Fit a Gaussian mixture models with a Dirichlet process prior
-# with ten components
-dpgmm = mixture.DirichletGaussianMixture(
+dpgmm = mixture.BayesianGaussianMixture(
     n_components=10, covariance_type='full', weight_concentration_prior=1e-2,
+    weight_concentration_prior_type='dirichlet_process',
     mean_precision_prior=1e-2, covariance_prior=1e0 * np.eye(2),
     init_params="random", max_iter=100, random_state=2).fit(X)
 plot_results(X, dpgmm.predict(X), dpgmm.means_, dpgmm.covariances_, 1,
-             r'Dirichlet Process with $\beta_0=0.01$')
+             "Bayesian Gaussian mixture models with a Dirichlet process prior "
+             r"for $\gamma_0=0.01$.")
 
-# Sample the Dirichlet process Gaussian mixture
-X_s, y_s = dpgmm.sample(n_samples=5000)
+X_s, y_s = dpgmm.sample(n_samples=2000)
 plot_samples(X_s, y_s, dpgmm.n_components, 0,
-             "Sampling of model obtained with a Dirichlet Process "
-             r"with $\beta_0=0.01$ and $5000$ samples")
+             "Gaussian mixture with a Dirichlet process prior "
+             r"for $\gamma_0=0.01$ sampled with $2000$ samples.")
 
-# Fit a Gaussian mixture models with a Dirichlet process prior
-# and ten components
-dpgmm = mixture.DirichletGaussianMixture(
+dpgmm = mixture.BayesianGaussianMixture(
     n_components=10, covariance_type='full', weight_concentration_prior=1e+2,
+    weight_concentration_prior_type='dirichlet_process',
     mean_precision_prior=1e-2, covariance_prior=1e0 * np.eye(2),
     init_params="kmeans", max_iter=100, random_state=2).fit(X)
 plot_results(X, dpgmm.predict(X), dpgmm.means_, dpgmm.covariances_, 2,
-             r'Dirichlet Process with $\beta_0=100$')
+             "Bayesian Gaussian mixture models with a Dirichlet process prior "
+             r"for $\gamma_0=100$")
 
-# Sample the Dirichlet process Gaussian mixture
-X_s, y_s = dpgmm.sample(n_samples=5000)
-plot_samples(X_s, y_s, dpgmm.n_components, 1, "Sampling of model obtained "
-             r"with a Dirichlet Process with $\beta_0=100$ and $5000$ samples")
+X_s, y_s = dpgmm.sample(n_samples=2000)
+plot_samples(X_s, y_s, dpgmm.n_components, 1,
+             "Gaussian mixture with a Dirichlet process prior "
+             r"for $\gamma_0=100$ sampled with $2000$ samples.")
 
 plt.show()